Keep a maximum-weight hyperedge from emptying the bisection
The hypergraph bisector's two restart loops started their running best at
u32::MAX, the largest cut a hypergraph can have: Hypergraph::new accepts a
total hyperedge weight of exactly u32::MAX, and a candidate that ties the
starting value is never taken. Every restart then left the partition empty,
and the balance pass indexed it and panicked.

Both loops now hold the best cut in u64 and start from u64::MAX, as the graph
bisector already does. Identical for every cut below the maximum.

+++ src/partition/hypergraph/mod.rs
@@ -226,7 +226,8 @@ pub fn multilevel_hypergraph_bisect(
     }
 
     let mut best_part = Vec::new();
-    let mut best_cut = u32::MAX;
+    // Wider than the cut it holds; see the same value in `initial`.
+    let mut best_cut = u64::MAX;
 
     // Best-of-N on the cut, the objective the caller asked for here; see "Where
     // the two bisectors differ" in the shared partition bookkeeping.
@@ -241,7 +242,7 @@ pub fn multilevel_hypergraph_bisect(
             config.effort,
             &mut scratch,
         );
-        let candidate_cut = hyperedge_cut(hg, &part);
+        let candidate_cut = u64::from(hyperedge_cut(hg, &part));
         if candidate_cut < best_cut {
             best_cut = candidate_cut;
             best_part = part;
